refactor firebase providers concealing firebaseConfig in .env
