--- frontend/src/app/admin/util/firebaseconfig.ts
@@ -1,8 +1,0 @@
-export const firebaseConfig = {
-  apiKey: "AIzaSyC0JaYSDXMzWivnh08itviRohomG6R_MHo",
-  authDomain: "flot-dev.firebaseapp.com",
-  projectId: "flot-dev",
-  storageBucket: "flot-dev.appspot.com",
-  messagingSenderId: "738782826196",
-  appId: "1:738782826196:web:e020d0f0bd0d98aeb8f541",
-};
